+++ modulo3/labex/projeto-labex/public/manifest.json
