+++ modulo3/labex/projeto-labex/src/App.js
@@ -1,0 +1,13 @@
+import React from "react";
+import {Router} from "./Routes/Router";
+
+function App() {
+  return (
+    <div>
+      <Router />
+      
+    </div>
+  )
+}
+
+export default App;
