+++ modulo3/labex/projeto-labex/src/pages/HomePage.js
@@ -1,0 +1,22 @@
+import React from "react"
+import { useNavigate } from "react-router-dom"
+
+const HomePage = () => {
+    const navigate = useNavigate()
+
+    const goToListTripsPage = () => {
+        navigate("/ListTrips")
+    }
+
+    const goToAdminHomePage = () => {
+        navigate("/AdminHome")
+    }
+return (
+   
+
+    <button onClick={goToListTripsPage}>Ver Viagens</button>
+
+)
+}
+
+export default HomePage;
